Refactor "More" bottom sheets into a set of navigable screens.
Precursor to https://github.com/bithyve/hexa/issues/1702.
