+++ src/components/AppBottomSheetTouchableWrapper.tsx
@@ -2,6 +2,7 @@ import { TouchableOpacity, TouchableOpacityProps, Platform} from "react-native";
 import { TouchableOpacity as TouchableOpacityGestureHandler } from "react-native-gesture-handler";
 import React, { ReactNode } from "react";
 
+
 export const AppBottomSheetTouchableWrapper = (
     props: { children: ReactNode } & Pick<TouchableOpacityProps,'onPress' | 'style' | 'activeOpacity' | 'disabled' | 'hitSlop'>,
   ) =>
