+++ src/pages/Home/Home.tsx
@@ -175,7 +175,6 @@ interface HomeStateTypes {
   selectedBottomTab: BottomTab;
   transactions: any[];
   modalData: any;
-  knowMoreBottomSheetsFlag: boolean;
   tabBarIndex: number;
   bottomSheetState: BottomSheetState;
   loading: boolean;
@@ -258,12 +257,10 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
   transactionTabBarBottomSheetRef = createRef<RNBottomSheet>();
   qrTabBarBottomSheetRef = createRef<RNBottomSheet>();
   addTabBarBottomSheetRef = createRef<RNBottomSheet>();
-  moreTabBarBottomSheetRef = createRef<RNBottomSheet>();
 
   // TODO: Completely replace `BottomSheet` with `RNBottomSheet` a la the refs above (https://trello.com/c/boUNRk6t)
   trustedContactRequestBottomSheetRef = createRef<BottomSheet>();
   transactionDetailsBottomSheetRef = createRef<BottomSheet>();
-  settingsBottomSheetRef = createRef<BottomSheet>();
   custodianRequestBottomSheetRef = createRef<BottomSheet>();
   errorBottomSheetRef = createRef<BottomSheet>();
   addContactAddressBookBottomSheetRef = createRef<BottomSheet>();
@@ -289,7 +286,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
       balances: {},
       selectedBottomTab: BottomTab.Transactions,
       transactions: [],
-      knowMoreBottomSheetsFlag: false,
       modalData: initialTransactionData,
       tabBarIndex: 999,
       bottomSheetState: BottomSheetState.Closed,
@@ -1060,7 +1056,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
           }
 
           setTimeout(() => {
-            this.settingsBottomSheetRef.current?.snapTo(0);
             this.custodianRequestBottomSheetRef.current?.snapTo(1);
             this.transactionTabBarBottomSheetRef.current?.snapTo(1);
           }, 2);
@@ -1147,7 +1142,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
               });
             }
             setTimeout(() => {
-              this.settingsBottomSheetRef.current?.snapTo(0);
               this.custodianRequestBottomSheetRef.current?.snapTo(1);
               this.transactionTabBarBottomSheetRef.current?.snapTo(1);
             }, 2);
@@ -1275,7 +1269,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
       this.transactionTabBarBottomSheetRef.current?.snapTo(1);
       this.addTabBarBottomSheetRef.current?.close();
       this.qrTabBarBottomSheetRef.current?.close();
-      this.moreTabBarBottomSheetRef.current?.close();
     }, 500);
 
     this.getAssociatedContact();
@@ -1683,63 +1676,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
       transactions: accumulativeTransactions,
       isBalanceLoading: false,
     });
-
-    // if (balancesParam) {
-    //   if (
-    //     JSON.stringify(balancesParam) !==
-    //     JSON.stringify({
-    //       testBalance,
-    //       regularBalance,
-    //       secureBalance,
-    //       accumulativeBalance,
-    //     })
-    //   ) {
-    //     setBalances({
-    //       testBalance,
-    //       regularBalance,
-    //       secureBalance,
-    //       accumulativeBalance,
-    //     });
-    //     setTransactions(accumulativeTransactions);
-    //   }
-    // } else {
-    //   setBalances({
-    //     testBalance,
-    //     regularBalance,
-    //     secureBalance,
-    //     accumulativeBalance,
-    //   });
-    //   setTransactions(accumulativeTransactions);
-    // }
-  };
-
-  onPressSettingsElements = async (type, currencycode) => {
-    const { navigation, currencyCode } = this.props;
-    if (type == 'ManagePin') {
-      return navigation.navigate('SettingManagePin', {
-        managePinSuccessProceed: (pin) => this.managePinSuccessProceed(pin),
-      });
-    } else if (type == 'ChangeCurrency') {
-      let currency = currencyCode;
-      //await AsyncStorage.getItem('currencyCode');
-      navigation.navigate('ChangeCurrency');
-      this.setState({
-        currencyCode: currency,
-      });
-    } else if (type == 'ChangeWalletName') {
-      navigation.navigate('SettingWalletNameChange');
-    }
-  };
-
-  managePinSuccessProceed = (pin) => {
-    this.setState(
-      {
-        tabBarIndex: 999,
-      },
-      () => {
-        this.settingsBottomSheetRef.current?.snapTo(0);
-      },
-    );
   };
 
   onNotificationListOpen = async () => {
@@ -1808,27 +1744,12 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
   onPhoneNumberChange = () => {};
 
   handleBottomTabSelection = (tab: BottomTab) => {
-    const knowMoreBottomSheetsFlag = tab === BottomTab.More;
-
     if (tab === BottomTab.More) {
-      this.setState(
-        {
-          knowMoreBottomSheetsFlag,
-          selectedBottomTab: tab,
-          bottomSheetState: BottomSheetState.Open,
-        },
-        () => {
-          this.transactionTabBarBottomSheetRef.current?.close();
-          this.addTabBarBottomSheetRef.current?.close();
-          this.qrTabBarBottomSheetRef.current?.close();
-          this.moreTabBarBottomSheetRef.current?.expand();
-        },
-      );
+      this.props.navigation.navigate('MoreOptions');
     }
     if (tab === BottomTab.Transactions) {
       this.setState(
         {
-          knowMoreBottomSheetsFlag,
           modalData: initialTransactionData,
           selectedBottomTab: tab,
           bottomSheetState: BottomSheetState.Open,
@@ -1837,14 +1758,12 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
           this.transactionTabBarBottomSheetRef.current?.expand();
           this.addTabBarBottomSheetRef.current?.close();
           this.qrTabBarBottomSheetRef.current?.close();
-          this.moreTabBarBottomSheetRef.current?.close();
         },
       );
     }
     if (tab === BottomTab.Add) {
       this.setState(
         {
-          knowMoreBottomSheetsFlag,
           modalData: [],
           selectedBottomTab: tab,
           bottomSheetState: BottomSheetState.Open,
@@ -1853,13 +1772,11 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
           this.transactionTabBarBottomSheetRef.current?.close();
           this.addTabBarBottomSheetRef.current?.expand();
           this.qrTabBarBottomSheetRef.current?.close();
-          this.moreTabBarBottomSheetRef.current?.close();
         },
       );
     } else if (tab === BottomTab.QR) {
       this.setState(
         {
-          knowMoreBottomSheetsFlag,
           modalData: initialTransactionData,
           selectedBottomTab: tab,
           bottomSheetState: BottomSheetState.Open,
@@ -1868,7 +1785,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
           this.transactionTabBarBottomSheetRef.current?.close();
           this.addTabBarBottomSheetRef.current?.close();
           this.qrTabBarBottomSheetRef.current?.expand();
-          this.moreTabBarBottomSheetRef.current?.close();
         },
       );
     }
@@ -2083,8 +1999,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
         return this.addTabBarBottomSheetRef;
       case BottomTab.QR:
         return this.qrTabBarBottomSheetRef;
-      case BottomTab.More:
-        return this.moreTabBarBottomSheetRef;
       default:
         return this.transactionTabBarBottomSheetRef;
     }
@@ -2146,7 +2060,7 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
     });
 
     if (value.info.includes('Trusted Contact request accepted by')) {
-      navigation.navigate('AddressBookContents');
+      navigation.navigate('FriendsAndFamily');
       return;
     }
 
@@ -2175,35 +2089,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
         this.notificationsListBottomSheetRef.current?.snapTo(0);
       }, 2);
       this.handleBottomTabSelection(BottomTab.Transactions);
-    }
-  };
-
-  onPressElement = (item) => {
-    const { navigation } = this.props;
-    if (item.title == 'Backup Health') {
-      navigation.navigate('ManageBackup');
-      return;
-    }
-    if (item.title == 'Friends and Family') {
-      navigation.navigate('AddressBookContents');
-      return;
-    } else if (item.title == 'Wallet Settings') {
-      this.settingsBottomSheetRef.current?.snapTo(1);
-      setTimeout(() => {
-        this.setState({
-          tabBarIndex: 0,
-        });
-      }, 10);
-    } else if (item.title == 'Funding Sources') {
-      navigation.navigate('ExistingSavingMethods');
-    } else if (item.title === 'Hexa Community (Telegram)') {
-      let url = 'https://t.me/HexaWallet';
-      Linking.openURL(url)
-        .then((data) => {})
-        .catch((e) => {
-          alert('Make sure Telegram installed on your device');
-        });
-      return;
     }
   };
 
@@ -2322,7 +2207,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
       selectedTransactionItem,
       tabBarIndex,
       deepLinkModalOpen,
-      knowMoreBottomSheetsFlag,
       errorMessageHeader,
       errorMessage,
       buttonText,
@@ -2567,41 +2451,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
                     scanedCode: this.processQRData,
                   });
                 }}
-              />
-            </BottomSheetView>
-          </RNBottomSheet>
-        )}
-
-        {!isLoading && (
-          <RNBottomSheet
-            ref={this.moreTabBarBottomSheetRef}
-            snapPoints={[
-              -50,
-              Platform.OS == 'ios' && DeviceInfo.hasNotch()
-                ? hp('18%')
-                : Platform.OS == 'android'
-                ? hp('19%')
-                : hp('18%'),
-              Platform.OS == 'ios' && DeviceInfo.hasNotch()
-                ? hp('65%')
-                : hp('64%'),
-            ]}
-            handleComponent={BottomSheetHandle}
-            onChange={(newPositionIndex: number) => {
-              this.handleBottomSheetPositionChange(
-                this.moreTabBarBottomSheetRef,
-                newPositionIndex,
-              );
-            }}
-          >
-            <BottomSheetView>
-              <BottomSheetHeader
-                title="More"
-                onPress={this.handleBottomSheetHeaderTap}
-              />
-
-              <MoreHomePageTabContents
-                onPressElements={(item) => this.onPressElement(item)}
               />
             </BottomSheetView>
           </RNBottomSheet>
@@ -2883,65 +2732,6 @@ class Home extends PureComponent<HomePropsTypes, HomeStateTypes> {
             )}
           />
         )}
-
-        {knowMoreBottomSheetsFlag ? (
-          <BottomSheet
-            onOpenEnd={() => {
-              if (!deepLinkModalOpen) {
-                this.setState({
-                  tabBarIndex: 0,
-                });
-              }
-            }}
-            onCloseEnd={() => {
-              if (!deepLinkModalOpen) {
-                this.setState({
-                  tabBarIndex: 999,
-                });
-              }
-            }}
-            enabledInnerScrolling={true}
-            ref={this.settingsBottomSheetRef}
-            snapPoints={[
-              -50,
-              Platform.OS == 'ios' && DeviceInfo.hasNotch()
-                ? hp('65%')
-                : hp('64%'),
-            ]}
-            renderContent={() => (
-              <SettingsContents
-                currencyCode={CurrencyCode}
-                onPressManagePin={(type, currencycode) =>
-                  this.onPressSettingsElements(type, currencycode)
-                }
-                onPressBack={() => {
-                  this.setState(
-                    {
-                      tabBarIndex: 999,
-                    },
-                    () => {
-                      this.settingsBottomSheetRef.current?.snapTo(0);
-                    },
-                  );
-                }}
-              />
-            )}
-            renderHeader={() => (
-              <SmallHeaderModal
-                borderColor={Colors.white}
-                backgroundColor={Colors.white}
-                onPressHeader={() => {
-                  this.setState(
-                    {
-                      tabBarIndex: 999,
-                    },
-                    () => this.settingsBottomSheetRef.current?.snapTo(0),
-                  );
-                }}
-              />
-            )}
-          />
-        ) : null}
 
         <BottomSheet
           onOpenEnd={() => {
